+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -336,7 +336,7 @@ describe('Datagrid filter tests', () => {
   }
 });
 
-fdescribe('Datagrid filter RTL tests', () => { //eslint-disable-line
+describe('Datagrid filter RTL tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-filter?locale=ar-SA&layout=nofrills');