+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # What's New with Enterprise
 
+## v4.27.0
+
+### v4.27.0 Fixes
+
+- `[Datagrid]` Fixed a bug were datagrid tree would have very big text in the tree nodes on IOS. ([#3347](https://github.com/infor-design/enterprise/issues/3347))
+
 ## v4.26.0
 
 ### v4.26.0 Features